+++ src/main/java/csc1035/project2/MainIO.java
@@ -11,50 +11,9 @@ import java.util.*;
  */
 public class MainIO {
     public static void main(String[] args) {
-        //addDummyData();
         new MainIO().run();
     }
 
-    public static void addDummyData() {
-        IController controller;
-        Modules[] modules = {new Modules("RSI3393","Nunc nisl.",20,12,new HashSet<>()),
-                new Modules("BBU5808","Integer pede justo lacinia eget tincidunt eget tempus vel pede.",10,12,new HashSet<>()),
-                new Modules("SAM0176","Cras in purus eu magna vulputate luctus.",10,6,new HashSet<>())};
-
-        Room[] rooms = {new Room("0.379","PC Cluster",157,36,new ArrayList<Booking>()),
-                new Room("0.365","PC Cluster",59,9,new ArrayList<Booking>()),
-                new Room("1.846","Lecture Lecture",274,50,new ArrayList<Booking>())};
-
-        Staff[] staff = {new Staff("NUC3292317","Jacenta","Khomich",new HashSet<>()),
-                new Staff("NUC9674902","Carmita","Cogzell",new HashSet<>()),
-                new Staff("NUC7362101","Swen","Geard",new HashSet<>())};
-
-        Students[] students = {new Students("216906208","Doralynn","Bordman",new HashSet<>()),
-                new Students("218577635","Farra","Pietroni", new HashSet<>()),
-                new Students("216365117","Clareta","Osmint",new HashSet<>())};
-
-        ModuleRequirements[] requirements = {new ModuleRequirements("RSI3393",1,2,1,2,new Date()),
-                new ModuleRequirements("BBU5808",1,2,2,2,new Date()),
-                new ModuleRequirements("SAM0176",3,1,4,1,new Date())};
-
-
-        controller = new Controller();
-
-        //controller.deleteAll(Take.class);
-        controller.deleteAll(Modules.class);
-        controller.deleteAll(Room.class);
-        controller.deleteAll(Staff.class);
-        controller.deleteAll(Students.class);
-        controller.deleteAll(ModuleRequirements.class);
-        controller.deleteAll(ModuleBooking.class);
-
-
-        controller.bulkListSave(Arrays.asList(modules));
-        controller.bulkListSave(Arrays.asList(rooms));
-        controller.bulkListSave(Arrays.asList(staff));
-        controller.bulkListSave(Arrays.asList(students));
-        controller.bulkListSave(Arrays.asList(requirements));
-    }
     public void run() {
 
         boolean quit = false;